fix hooks and filter func
